# Back/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/enroll:
    post:
      summary: Payment
      operationId: payment
      description: 'Payment Process'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Enrollment successful'
                properties:
                  message:
                    type: string
                    example: 'Enrollment successful'
        500:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: "No such PaymentMethod: 'pm_1NS6eqGEPWFHPllANwhUzh'"
                properties:
                  message:
                    type: string
                    example: "No such PaymentMethod: 'pm_1NS6eqGEPWFHPllANwhUzh'"
      tags:
        - 'Enrollment And Payment'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                course_ids:
                  type: object
                  description: 'Array of courses ids.'
                  example:
                    - 1
                    - 10
                  properties: []
                payment_amount:
                  type: number
                  description: 'Amount needed to pay.'
                  example: 100.0
                payment_id:
                  type: string
                  description: 'Payment method id that generated from stripe. Example ='
                  example: nihil
              required:
                - course_ids
                - payment_amount
                - payment_id
...
  '/api/lectures/{id}':
    get:
      summary: 'Display Lecture'
      operationId: displayLecture
      description: 'Display the specified resource.'
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Lectures
    put:
      summary: 'Update Lecture'
      operationId: updateLecture
      description: 'Update Lecture in course.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    id: 2
                    title: 'new lec edited'
                    type: video
                    content: /storage/content/courses/course1/section1/lecture2/lecture.mkv
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 2
                      title:
                        type: string
                        example: 'new lec edited'
                      type:
                        type: string
                        example: video
                      content:
                        type: string
                        example: /storage/content/courses/course1/section1/lecture2/lecture.mkv
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'The title field is required. (and 1 more error)'
                  errors:
                    title:
                      - 'The title field is required.'
                    type:
                      - 'The type field is required.'
                properties:
                  message:
                    type: string
                    example: 'The title field is required. (and 1 more error)'
                  errors:
                    type: object
                    properties:
                      title:
                        type: array
                        example:
                          - 'The title field is required.'
                        items:
                          type: string
                      type:
                        type: array
                        example:
                          - 'The type field is required.'
                        items:
                          type: string
      tags:
        - Lectures
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 'Title of lecture.'
                  example: 'lec 1'
                type:
                  type: string
                  description: '[video or text].'
                  example: video
                content:
                  type: string
                  format: binary
                  description: 'Content of lecture'
              required:
                - title
                - type
    delete:
      summary: 'Delete Lecture'
      operationId: deleteLecture
      description: 'Remove Lecture from course.'
      parameters: []
      responses:
        204:
          description: ''
      tags:
        - Lectures
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the lecture.'
        example: 9
        required: true
        schema:
          type: integer
...